requirements.txt
fastapi
uvicorn
torch
kafka-python
feast
pymilvus
evidently
prometheus-client
